Level-set jobs, organizations, and the inference API to v2
JobsManager's three targetType class attributes now hold the v2 vocabulary
('Cluster' / 'VirtualCluster', no legacy self-managed target), and the new
v1/job.py subclass overrides them back to v1's.

That flip is only safe if v1 organizations keep handing out the v1
JobsManager, so v1/organization.py becomes a real subclass repointing
_jobs_manager_class -- otherwise v1 job scheduling would silently start
sending v2 targetType values. Organizations.current hardcoded the base
Organization, so which class it hands out is now a class attribute the v1
subclass repoints too.

The inference API exists at v1 only, so its implementation moves into
v1/inference_api.py and the raising v2 subclass is deleted; the shared
Organization holds _inference_api_manager_class = None and raises from
inference_apis. Top-level inference_api.py becomes a shim onto v1, matching
export.py, so Fusion's imports are untouched.

Note: 'v2 has no inference routes' is inferred from the raising subclass this
commit deletes, not confirmed against the live API.
